requirements_dev.txt
-e .
coverage==7.12.0
mypy==1.19.0
pdoc==16.0.0
pip==25.3
pre-commit==4.5.0
pytest-cov==7.0.0
pytest==9.0.1
ruff==0.14.7

aiohttp==3.13.2
mashumaro==3.17
pytest-aiohttp==1.1.0

# For user management CLI
pyyaml>=6.0
PyJWT==2.10.1
